+++ tedge/src/cli/connect/test.rs
@@ -1,4 +1,6 @@
-use super::*;
+use crate::cli::connect::*;
+use crate::config::TEdgeConfig;
+use tempfile::NamedTempFile;
 
 const CORRECT_URL: &str = "http://test.com";
 const INCORRECT_URL: &str = "noturl";
